refactor(agents): split isolation support modules
Move isolation validation, spawn setup, runtime ownership, and test fixtures into focused sibling modules so existing Rust files satisfy the size ratchet without changing behavior.
