+++ desktop/src-tauri/src/commands/agents.rs
@@ -577,20 +577,7 @@ pub async fn create_managed_agent(
         .map(str::trim)
         .filter(|value| !value.is_empty())
         .map(str::to_string);
-    if let Some(parallelism) = input.parallelism {
-        if !(1..=32).contains(&parallelism) {
-            return Err("parallelism must be between 1 and 32".to_string());
-        }
-    }
-    crate::managed_agents::validate_user_env_keys(&input.env_vars)?;
-    if let Some(profile) = &input.filesystem_isolation {
-        crate::managed_agents::validate_filesystem_isolation_profile(profile)?;
-        if input.backend != BackendKind::Local {
-            return Err(
-                "filesystem isolation is available only for local managed agents".to_string(),
-            );
-        }
-    }
+    super::agent_fs::validate_create(&input)?;
 
     // Validate & normalize the respond-to allowlist BEFORE any side effects.
     // The harness has its own validator (buzz-acp/src/config.rs) but we want
